rag-policy-assistant/requirements.txt
streamlit
langchain-core
langchain-groq
python-dotenv
pypdf
chromadb
